% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_types.R
\name{apply_types}
\alias{apply_types}
\title{Convert column-type according to column_name}
\usage{
apply_types(data, timestamp_to_date = TRUE)
}
\arguments{
\item{data}{tibble}

\item{timestamp_to_date}{boolean convert timestamp to date}
}
\value{
tibble
}
\description{
Convert column-type according to column_name
}
