% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_app_detail.R
\name{get_app_detail}
\alias{get_app_detail}
\title{get_app_detail fetching details about an app}
\usage{
get_app_detail(token, app)
}
\arguments{
\item{token}{login token generated by \code{\link{get_login_token}}}

\item{app}{app id, see \code{\link{get_apps_list}}}
}
\value{
data.frame with details about the app
}
\description{
get_app_detail fetching details about an app
}
