% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_apps_list.R
\name{get_apps_list_single}
\alias{get_apps_list_single}
\title{single request for getting apps}
\usage{
get_apps_list_single(token, offset = 0)
}
\arguments{
\item{token}{login token generated by \code{\link{get_login_token}}}

\item{offset}{defailt 0, integer, where to start fetching apps}
}
\value{
data.frame
}
\description{
single request for getting apps
}
