% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_generic_list.R
\name{get_generic_list}
\alias{get_generic_list}
\title{get_generic_list fetches all rows of a given api}
\usage{
get_generic_list(token, offset, api_endpoint, columns, datetime_columns)
}
\arguments{
\item{token}{login token generated by \code{\link{get_login_token}}}

\item{offset}{defailt 0, integer, where to start fetching usergroups}

\item{api_endpoint}{string with api endpoint}

\item{columns}{vector of strings containing columnnames}

\item{datetime_columns}{vector of strings containing columnnames which are
converted to datetime}
}
\value{
data.frame
}
\description{
get_generic_list fetches all rows of a given api
}
