% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_login_token.R
\name{get_login_token}
\alias{get_login_token}
\title{Fetch login token}
\usage{
get_login_token(credentials)
}
\arguments{
\item{credentials}{List with fields client_id, client_secret and url}
}
\value{
List with login_token
}
\description{
Fetch login token
}
