% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_apps_list.R
\name{get_permissions_for_app}
\alias{get_permissions_for_app}
\title{get_permissions_for_app}
\usage{
get_permissions_for_app(token, app)
}
\arguments{
\item{token}{login token generated by \code{\link{get_login_token}}}

\item{app}{app id, see \code{\link{get_apps_list}}}
}
\value{
data.frame
}
\description{
get_permissions_for_app
}
