% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_apps_list.R
\name{get_permissions_for_metasite}
\alias{get_permissions_for_metasite}
\title{get_permissions_for_metasite}
\usage{
get_permissions_for_metasite(token, metasite)
}
\arguments{
\item{token}{login token generated by \code{\link{get_login_token}}}

\item{metasite}{app id, see \code{\link{get_metasites_list}}}
}
\value{
data.frame
}
\description{
get_permissions_for_metasite
}
