% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_test_credentials}
\alias{get_test_credentials}
\title{get_test_credentials
getting credentials for testing}
\usage{
get_test_credentials()
}
\value{
list
}
\description{
Fills credentials out of ENV into a list
}
