% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_usergroups_list.R
\name{get_usergroup_members}
\alias{get_usergroup_members}
\title{get_usergroup_members}
\usage{
get_usergroup_members(token, usergroup_id)
}
\arguments{
\item{token}{login token generated by \code{\link{get_login_token}}}

\item{usergroup_id}{usergroup (see \code{\link{get_usergroups_list}})}
}
\value{
data.frame
}
\description{
get_usergroup_members
}
