% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_usergroups_list.R
\name{get_usergroups_list}
\alias{get_usergroups_list}
\title{get_usergroups_list fetch all usergroups}
\usage{
get_usergroups_list(token, offset = 0)
}
\arguments{
\item{token}{login token generated by \code{\link{get_login_token}}}

\item{offset}{defailt 0, integer, where to start fetching usergroups}
}
\value{
data.frame
}
\description{
get_usergroups_list fetch all usergroups
}
