% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_users_list.R
\name{get_users_list}
\alias{get_users_list}
\title{get_users_list fetch all users}
\usage{
get_users_list(token, offset = 0)
}
\arguments{
\item{token}{login token generated by \code{\link{get_login_token}}}

\item{offset}{defailt 0, integer, where to start fetching users}
}
\value{
data.frame
}
\description{
get_users_list fetch all users
}
