% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_column_a_metric.R
\name{is_column_a_metric}
\alias{is_column_a_metric}
\title{is_column_a_metric
Checks if column_name indicates numeric values
Uses https://developers.piwik.pro/en/latest/custom_reports/columns.html}
\usage{
is_column_a_metric(column_name)
}
\arguments{
\item{column_name}{string}
}
\value{
boolean
}
\description{
is_column_a_metric
Checks if column_name indicates numeric values
Uses https://developers.piwik.pro/en/latest/custom_reports/columns.html
}
