% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{MAX_LINES_PER_REQUEST_USERS_API}
\alias{MAX_LINES_PER_REQUEST_USERS_API}
\title{Maximum number of lines requested in Users-API}
\usage{
MAX_LINES_PER_REQUEST_USERS_API()
}
\value{
int
}
\description{
Maximum number of lines requested in Users-API
}
