% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cran_release_questions.R
\name{release_questions}
\alias{release_questions}
\title{Ask package specific question during release-process}
\usage{
release_questions()
}
\value{
vector of strings
}
\description{
Ask package specific question during release-process
}
