% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send_query.R
\name{send_query}
\alias{send_query}
\title{Send the query and receive the result}
\usage{
send_query(
  query,
  token,
  use_csv = TRUE,
  fetch_by_day = FALSE,
  api = "query",
  caching = FALSE,
  caching_dir = "cache",
  convert_types = TRUE
)
}
\arguments{
\item{query}{list generated by \code{\link{build_query}}}

\item{token}{login token}

\item{use_csv}{logical to choose whether to fetch data via extra csv-request}

\item{fetch_by_day}{logical fetch data day by day}

\item{api}{API endpoint ("query", "sessions", "events")}

\item{caching}{logical Set TRUE to enable caching}

\item{caching_dir}{character Set directory for saving caching data, default cache}

\item{convert_types}{logical guess type of columns and set them}
}
\value{
result as tibble
}
\description{
Send the query and receive the result
}
