% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send_query.R
\name{send_query_single}
\alias{send_query_single}
\title{Send the query and receive the result}
\usage{
send_query_single(query, token, use_csv, api, caching, caching_dir)
}
\arguments{
\item{query}{list generated by build_query()}

\item{token}{login token}

\item{use_csv}{logical to choose whether to fetch data via extra csv-request}

\item{api}{API endpoint (query, sessions, events)}

\item{caching}{logical Set TRUE to enable caching}

\item{caching_dir}{character Set directory for saving caching data}
}
\value{
result as list with values data and meta
}
\description{
Send the query and receive the result
}
