% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_types.R
\name{apply_types}
\alias{apply_types}
\title{Convert column-type according to column_name}
\usage{
apply_types(data)
}
\arguments{
\item{data}{tibble}
}
\value{
tibble
}
\description{
Convert column-type according to column_name
}
