% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_apps_list.R
\name{get_apps_list}
\alias{get_apps_list}
\title{get_apps_list fetch all apps}
\usage{
get_apps_list(token)
}
\arguments{
\item{token}{login token generated by \code{\link{get_login_token}}}
}
\value{
data.frame
}
\description{
get_apps_list fetch all apps
}
