% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_types.R
\name{get_column_type}
\alias{get_column_type}
\title{Convert column-type according to column_name}
\usage{
get_column_type(column_name)
}
\arguments{
\item{column_name}{string}
}
\value{
string suggested type of column
}
\description{
Convert column-type according to column_name
}
